# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: trapline API
  version: "1.0"
  summary: Errors, releases, alerts, monitors, tracing and release health for one installation.
  description: |
    One API, four clients: this REST surface, the `trapline` CLI, the web
    panel, and MCP (ADR 006). Everything the panel can do is reachable here,
    because the panel is a client of this and not a privileged insider.

    **Authentication.** Two credentials, one gate. A browser session carries a
    `SameSite=Lax` cookie and must send `X-Trapline-Request` on every
    state-changing request (the custom-header CSRF defence). A program carries
    `Authorization: Bearer ek_…` and sends no such header, because a bearer
    token is not ambient and there is nothing to forge.

    **Scopes.** Coarse on purpose. `projects:read|write` covers issues, stats,
    releases, tracing, artefacts and health. `alerts:read|write` is separate
    because an alert channel holds a credential for somebody else's system.
    `monitors:read|write` is separate because a ping key is a credential, and
    because creating an uptime monitor is the permission to make this server
    open outbound connections to an address the caller chose (ADR 016).

    **Errors.** One shape for every failure: `{"error": "…"}`, with a status
    code that means what it says. A 400 from a search is a search that could
    not run, which is not the same as a search that found nothing (ADR 011).
  license:
    name: Apache-2.0
...
servers:
...
security:
  - sessionCookie: []
  - bearerToken: []
...
paths:
...
  /projects/{id}/issues/{issueID}/status:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/IssueID"
    post:
      tags: [issues]
      summary: Resolve, ignore or reopen.
      description: |
        `in_next_release` qualifies `resolved`: fixed, and the fix ships next,
        so events still arriving from the release it is being seen in are
        expected rather than proof the fix failed. Those events are counted
        and exposed as `seen_in_resolved_release_count`; they do not reopen
        the issue (ADR 012).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/IssueStatus" }
                in_next_release: { type: boolean, default: false }
      responses:
        "200":
          description: The status it now has.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { $ref: "#/components/schemas/IssueStatus" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /projects/{id}/stats:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/From"
      - $ref: "#/components/parameters/To"
    get:
      tags: [stats]
      summary: Events per hour, and the totals per level.
      responses:
        "200":
          description: The series.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectSeries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /projects/{id}/issues/{issueID}/stats:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/IssueID"
      - $ref: "#/components/parameters/Window"
    get:
      tags: [stats]
      summary: One issue's own history.
      responses:
        "200":
          description: The series.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssueSeries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --------------------------------------------------------------- releases
...
  /status/{slug}:
    servers:
      - url: /
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
        description: The project slug.
    get:
      tags: [ingest]
      summary: The public status page.
      description: |
        HTML rendered on the server, no JavaScript required, cached for 30
        seconds, rate-limited per address, and served only if the project has
        switched it on. The figures come from the daily aggregate and never
        from the individual checks (ADR 001, ADR 017). It is a page, not an
        endpoint: it is described here so that the route gate can see it, not
        because anything should parse it.
      security: []
      responses:
        "200":
          description: The page.
          content:
            text/html:
              schema: { type: string }
        "404": { description: No such project, or its status page is off. }
        "429": { description: Over the per-address limit. }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: trapline_session
      description: |
        The panel's credential. `SameSite=Lax`, `HttpOnly`, and `Secure` when
        the server knows it is behind TLS. Because it is ambient, every
        state-changing request must also carry `X-Trapline-Request` with any
        value; without it the request is refused with 403. A browser cannot
        set a custom header cross-origin without a preflight, and no
        permissive CORS policy is served for one to pass — which is the whole
        defence, in one header and no token plumbing.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: ek_...
      description: |
        A program's credential, carrying scopes. Not ambient, so it needs no
        CSRF header. Scopes are `projects:read`, `projects:write`,
        `alerts:read`, `alerts:write`, `monitors:read`, `monitors:write`;
        where an operation below names none, it takes `projects:read` or
        `projects:write` according to its method.
  parameters:
    ProjectID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
      description: The project's numeric id.
    IssueID:
      name: issueID
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    From:
      name: from
      in: query
      schema: { type: string, format: date-time }
      description: |
        The start of the range, RFC 3339. Absent means the endpoint's default
        window; the range actually used is echoed back in every response, so a
        client never has to guess what it got.
    To:
      name: to
      in: query
      schema: { type: string, format: date-time }
      description: The end of the range, RFC 3339.
    Window:
      name: range
      in: query
      schema: { type: string, enum: ["24h", "14d"], default: "24h" }
      description: |
        Two named windows rather than a duration: `24h` is the incident view,
        `14d` is "did the fix hold". A free-form duration invites `1000d`,
        which is a range query nobody meant.
...
  responses:
    BadRequest:
      description: |
        Malformed, or a question that cannot be asked. A search that the index
        cannot run answers here and says why — which is not the same as a
        search that found nothing, and the difference is an afternoon
        (ADR 011).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No credential, or one that has expired or been revoked.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        A valid credential without the scope this route demands — or, for a
        cookie, a state-changing request with no `X-Trapline-Request` header.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such resource. Also what a subsystem this build was assembled
        without answers: a route that does not exist is the honest description
        of a surface this installation does not have.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      description: The one failure shape, for every route in this document.
      required: [error]
      properties:
        error: { type: string }
...
    Range:
      type: object
      description: The window a response actually answered over, always echoed back.
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        hours: { type: integer }
    ProjectRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
...
    IssueStatus:
      type: string
      enum: [unresolved, resolved, ignored]
...
    HourPoint:
      type: object
      properties:
        hour: { type: string, description: 'UTC, `YYYY-MM-DDTHH`.' }
        count: { type: integer, format: int64 }
        by_level:
          type: object
          additionalProperties: { type: integer, format: int64 }
    ProjectSeries:
      type: object
      properties:
        project: { $ref: "#/components/schemas/ProjectRef" }
        range: { $ref: "#/components/schemas/Range" }
        total: { type: integer, format: int64 }
        series:
          type: array
          items: { $ref: "#/components/schemas/HourPoint" }
        by_level:
          type: object
          additionalProperties: { type: integer, format: int64 }
...
    IssueSeries:
      type: object
      properties:
        issue_id: { type: integer, format: int64 }
        range: { type: string, enum: ["24h", "14d"] }
        range_bounds: { $ref: "#/components/schemas/Range" }
        total: { type: integer, format: int64 }
        series:
          type: array
          items:
            type: object
            properties:
              hour: { type: string }
              count: { type: integer, format: int64 }
...